Refactored decorated marks into separate file
